fix : 리뷰 적용 (#375)
