+++ KCS/KCSUnitTest/TestDouble/UseCase/StubFailureStoreUpdateRequestUseCase.swift
@@ -1,27 +1,13 @@
 //
-//  StubStoreUpdateRequestUseCase.swift
+//  StubFailureStoreUpdateRequestUseCase.swift
 //  KCSUnitTest
 //
-//  Created by 조성민 on 3/5/24.
+//  Created by 조성민 on 3/6/24.
 //
 
 @testable import KCS
 import RxSwift
 import Alamofire
-
-struct StubSuccessStoreUpdateRequestUseCase: StoreUpdateRequestUseCase {
-    
-    let repository: StoreUpdateRequestRepository
-    
-    init(repository: StoreUpdateRequestRepository = StoreUpdateRequestRepositoryImpl(session: Session.default)) {
-        self.repository = repository
-    }
-    
-    func execute(type: StoreUpdateRequestType, storeID: Int, content: String) -> Observable<Void> {
-        return .just(())
-    }
-    
-}
 
 struct StubInternetFailureStoreUpdateRequestUseCase: StoreUpdateRequestUseCase {
     
@@ -76,22 +62,3 @@ struct StubClientFailureStoreUpdateRequestUseCase: StoreUpdateRequestUseCase {
     }
     
 }
-
-struct StubAlamofireFailureStoreUpdateRequestUseCase: StoreUpdateRequestUseCase {
-    
-    let repository: StoreUpdateRequestRepository
-    
-    init(repository: StoreUpdateRequestRepository = StoreUpdateRequestRepositoryImpl(session: Session.default)) {
-        self.repository = repository
-    }
-    
-    func execute(type: StoreUpdateRequestType, storeID: Int, content: String) -> Observable<Void> {
-        return Observable.create { observer -> Disposable in
-            observer.onError(ErrorAlertMessage.client)
-            
-            return Disposables.create()
-        }
-    }
-    
-}
-
